mailq: Add test for process invocation and graph definition

+++ mackerel-plugin-mailq/main.go
@@ -96,7 +96,7 @@ func (p *plugin) GraphDefinition() map[string]mp.Graphs {
 			Label: "Messages in Mail Queue",
 			Unit:  "integer",
 			Metrics: []mp.Metrics{
-				{Name: p.mta, Label: p.mta, Type: "uint32"},
+				{Name: p.mta, Label: p.mta, Type: "uint64"},
 			},
 		},
 	}
